% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aklimate-utils.R
\name{collapse_weights}
\alias{collapse_weights}
\title{Combine entries in a vector that match a particular pattern}
\usage{
collapse_weights(patterns, weights)
}
\arguments{
\item{patterns}{- vector of patterns to match}

\item{weights}{- named numeric vector that needs to be collapsed based on the patterns}
}
\value{
A named vector of scores for each pattern, sorted from largest to smallest.
}
\description{
Combine entries in a vector that match a particular pattern
}
