% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aklimate-utils.R
\name{rank.importance.type}
\alias{rank.importance.type}
\title{Relative contribution of data types based on cumulative importance of features}
\usage{
rank.importance.type(suffs, ranked, intron = "_")
}
\arguments{
\item{suffs}{- vector of suffixes of participating data types}

\item{ranked}{- vactor of ranked features produced by rank.features()}
}
\value{
A vector of controbutions for each data type, normalized to a sum of 1.
}
\description{
Relative contribution of data types based on cumulative importance of features
}
