% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aklimate-utils.R
\name{rank_features}
\alias{rank_features}
\title{rank_features}
\usage{
rank_features(akl_obj)
}
\arguments{
\item{akl_obj}{an aklimate model}
}
\value{
A ranked vector of all features with non-zero importance.
}
\description{
Feature importance calculation from an aklimate model
}
