% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aklimate.R
\name{predict.aklimate}
\alias{predict.aklimate}
\title{Compute predictions from an aklimate model}
\usage{
\method{predict}{aklimate}(akl_obj, dat, fsets, kernels = NULL,
  store.kernels = FALSE)
}
\arguments{
\item{akl_obj}{- an aklimate model}

\item{dat}{- samples x features data frame where columns might be of different type}

\item{fsets}{- list of prior knowledge feature sets}

\item{kernels}{- (Optional) pre-computed kernels for the MKL prediction part. By default the test kernels are computed on the fly from the RF models stored in the aklimate object.}
}
\value{
A vector of predictions.
}
\description{
Compute predictions from an aklimate model
}
